# Improve activity query performance in Constituent Detail Report