Overview
----------------------------------------
The Constituent Detail Report pulls all of the activities that are related to a contact, whether because the contact is the target, assignee, or source of the activity or because the activity is on a case involving the client.

The report had been joining the `civicrm_activity_contact` table three times, plus the case tables, and using `OR` clauses in the `WHERE` clause to filter the appropriate ones.  My understanding from @eileenmcnaughton  is that `OR` can be bad for performance, and it could be particularly severe in this case, as `civicrm_activity_contact` can be quite large.  It's also needless: the report typically is used for a single contact, and that contact's activities will be just a tiny part of the whole.

This changes the query to be run as a union of four subqueries, each with an inner join to the relevant contact-related table, with the contact ID specified in the join.  Consequently, each subquery should have a small number of results and should be relatively quick.

Before
----------------------------------------
The constituent detail report would take a long time or even crash MySQL if there were a lot of activities.

After
----------------------------------------
There should be no change in the result, but the query performs significantly faster.  On a site with 877,852 rows in `civicrm_activity_contact`, the query would crash the server.  Now, it runs in under 2 seconds.

Technical Details
----------------------------------------
Here's the old query that was formed:

```sql
SELECT activity_civireport.id                      AS civicrm_activity_id,
       activity_civireport.activity_type_id        AS
       civicrm_activity_activity_type_id,
       activity_civireport.subject                 AS civicrm_activity_subject,
       activity_civireport.activity_date_time      AS
       civicrm_activity_activity_date_time,
       activity_civireport.status_id               AS
       civicrm_activity_activity_status_id,
       activity_target_civireport.display_name     AS
       activity_target_civireport_target_contact_id,
       civicrm_activity_target.contact_id          AS target_contact_id,
       activity_assignment_civireport.display_name AS
       activity_assignment_civireport_assignee_contact_id,
       civicrm_activity_assignment.contact_id      AS assignee_contact_id,
       activity_source_civireport.display_name     AS
       activity_source_civireport_source_contact_id,
       civicrm_activity_source.contact_id,
       activity_source_civireport.display_name     AS added_by
FROM   civicrm_activity activity_civireport
       LEFT JOIN civicrm_activity_contact civicrm_activity_target
              ON activity_civireport.id = civicrm_activity_target.activity_id
                 AND civicrm_activity_target.record_type_id = 3
       LEFT JOIN civicrm_activity_contact civicrm_activity_assignment
              ON activity_civireport.id =
                 civicrm_activity_assignment.activity_id
                 AND civicrm_activity_assignment.record_type_id = 1
       LEFT JOIN civicrm_activity_contact civicrm_activity_source
              ON activity_civireport.id = civicrm_activity_source.activity_id
                 AND civicrm_activity_source.record_type_id = 2
       LEFT JOIN civicrm_contact activity_target_civireport
              ON civicrm_activity_target.contact_id =
                 activity_target_civireport.id
       LEFT JOIN civicrm_contact activity_assignment_civireport
              ON civicrm_activity_assignment.contact_id =
                 activity_assignment_civireport.id
       LEFT JOIN civicrm_contact activity_source_civireport
              ON civicrm_activity_source.contact_id =
                 activity_source_civireport.id
       LEFT JOIN civicrm_option_value
              ON ( activity_civireport.activity_type_id =
                   civicrm_option_value.value )
       LEFT JOIN civicrm_option_group
              ON civicrm_option_group.id = civicrm_option_value.option_group_id
       LEFT JOIN civicrm_case_activity
              ON civicrm_case_activity.activity_id = activity_civireport.id
       LEFT JOIN civicrm_case
              ON civicrm_case_activity.case_id = civicrm_case.id
       LEFT JOIN civicrm_case_contact
              ON civicrm_case_contact.case_id = civicrm_case.id
WHERE  ( civicrm_activity_source.contact_id IN ( 82, 141, 28, 9,
                                                 30, 94, 202, 146,
                                                 92, 158 )
          OR civicrm_activity_target.contact_id IN ( 82, 141, 28, 9,
                                                     30, 94, 202, 146,
                                                     92, 158 )
          OR civicrm_activity_assignment.contact_id IN ( 82, 141, 28, 9,
                                                         30, 94, 202, 146,
                                                         92, 158 )
          OR civicrm_case_contact.contact_id IN ( 82, 141, 28, 9,
                                                  30, 94, 202, 146,
                                                  92, 158 ) )
       AND civicrm_option_group.name = 'activity_type'
       AND activity_civireport.is_test = 0
       AND (( civicrm_option_value.component_id IS NULL
               OR civicrm_option_value.component_id IN ( 1, 2, 3, 4,
                                                         6, 8 ) ))
ORDER  BY activity_civireport.activity_date_time DESC   
```

The new query is below.  Note that the four subqueries are mostly the same; they just differ one where the inner join is.

```sql
(SELECT activity_civireport.id                      AS civicrm_activity_id,
        activity_civireport.activity_type_id        AS
        civicrm_activity_activity_type_id,
        activity_civireport.subject                 AS civicrm_activity_subject,
        activity_civireport.activity_date_time      AS
        civicrm_activity_activity_date_time,
        activity_civireport.status_id               AS
        civicrm_activity_activity_status_id,
        activity_target_civireport.display_name     AS
        activity_target_civireport_target_contact_id,
        civicrm_activity_target.contact_id          AS target_contact_id,
        activity_assignment_civireport.display_name AS
        activity_assignment_civireport_assignee_contact_id,
        civicrm_activity_assignment.contact_id      AS assignee_contact_id,
        activity_source_civireport.display_name     AS
        activity_source_civireport_source_contact_id,
        civicrm_activity_source.contact_id,
        activity_source_civireport.display_name     AS added_by,
        activity_civireport.activity_date_time      AS date_time_for_sort
 FROM   civicrm_activity activity_civireport
        JOIN civicrm_case_activity
          ON civicrm_case_activity.activity_id = activity_civireport.id
        JOIN civicrm_case
          ON civicrm_case_activity.case_id = civicrm_case.id
        JOIN civicrm_case_contact
          ON civicrm_case_contact.case_id = civicrm_case.id
             AND civicrm_case_contact.contact_id IN ( 82, 141, 28, 9,
                                                      30, 94, 202, 146,
                                                      92, 158 )
        LEFT JOIN civicrm_activity_contact civicrm_activity_assignment
               ON activity_civireport.id =
                  civicrm_activity_assignment.activity_id
                  AND civicrm_activity_assignment.record_type_id = 1
        LEFT JOIN civicrm_activity_contact civicrm_activity_source
               ON activity_civireport.id = civicrm_activity_source.activity_id
                  AND civicrm_activity_source.record_type_id = 2
        LEFT JOIN civicrm_activity_contact civicrm_activity_target
               ON activity_civireport.id = civicrm_activity_target.activity_id
                  AND civicrm_activity_target.record_type_id = 3
        LEFT JOIN civicrm_contact activity_assignment_civireport
               ON civicrm_activity_assignment.contact_id =
                  activity_assignment_civireport.id
        LEFT JOIN civicrm_contact activity_source_civireport
               ON civicrm_activity_source.contact_id =
                  activity_source_civireport.id
        LEFT JOIN civicrm_contact activity_target_civireport
               ON civicrm_activity_target.contact_id =
                  activity_target_civireport.id
        JOIN civicrm_option_value
          ON activity_civireport.activity_type_id = civicrm_option_value.value
             AND ( civicrm_option_value.component_id IS NULL
                    OR civicrm_option_value.component_id IN ( 1, 2, 3, 4,
                                                              6, 8 ) )
        JOIN civicrm_option_group
          ON civicrm_option_group.id = civicrm_option_value.option_group_id
             AND civicrm_option_group.name = 'activity_type'
 WHERE  activity_civireport.is_test = 0)
UNION
(SELECT activity_civireport.id                      AS civicrm_activity_id,
        activity_civireport.activity_type_id        AS
        civicrm_activity_activity_type_id,
        activity_civireport.subject                 AS civicrm_activity_subject,
        activity_civireport.activity_date_time      AS
        civicrm_activity_activity_date_time,
        activity_civireport.status_id               AS
        civicrm_activity_activity_status_id,
        activity_target_civireport.display_name     AS
        activity_target_civireport_target_contact_id,
        civicrm_activity_target.contact_id          AS target_contact_id,
        activity_assignment_civireport.display_name AS
        activity_assignment_civireport_assignee_contact_id,
        civicrm_activity_assignment.contact_id      AS assignee_contact_id,
        activity_source_civireport.display_name     AS
        activity_source_civireport_source_contact_id,
        civicrm_activity_source.contact_id,
        activity_source_civireport.display_name     AS added_by,
        activity_civireport.activity_date_time      AS date_time_for_sort
 FROM   civicrm_activity activity_civireport
        JOIN civicrm_activity_contact civicrm_activity_assignment
          ON activity_civireport.id = civicrm_activity_assignment.activity_id
             AND civicrm_activity_assignment.record_type_id = 1
             AND civicrm_activity_assignment.contact_id IN ( 82, 141, 28, 9,
                                                             30, 94, 202, 146,
                                                             92, 158 )
        LEFT JOIN civicrm_activity_contact civicrm_activity_source
               ON activity_civireport.id = civicrm_activity_source.activity_id
                  AND civicrm_activity_source.record_type_id = 2
        LEFT JOIN civicrm_activity_contact civicrm_activity_target
               ON activity_civireport.id = civicrm_activity_target.activity_id
                  AND civicrm_activity_target.record_type_id = 3
        LEFT JOIN civicrm_contact activity_assignment_civireport
               ON civicrm_activity_assignment.contact_id =
                  activity_assignment_civireport.id
        LEFT JOIN civicrm_contact activity_source_civireport
               ON civicrm_activity_source.contact_id =
                  activity_source_civireport.id
        LEFT JOIN civicrm_contact activity_target_civireport
               ON civicrm_activity_target.contact_id =
                  activity_target_civireport.id
        JOIN civicrm_option_value
          ON activity_civireport.activity_type_id = civicrm_option_value.value
             AND ( civicrm_option_value.component_id IS NULL
                    OR civicrm_option_value.component_id IN ( 1, 2, 3, 4,
                                                              6, 8 ) )
        JOIN civicrm_option_group
          ON civicrm_option_group.id = civicrm_option_value.option_group_id
             AND civicrm_option_group.name = 'activity_type'
 WHERE  activity_civireport.is_test = 0)
UNION
(SELECT activity_civireport.id                      AS civicrm_activity_id,
        activity_civireport.activity_type_id        AS
        civicrm_activity_activity_type_id,
        activity_civireport.subject                 AS civicrm_activity_subject,
        activity_civireport.activity_date_time      AS
        civicrm_activity_activity_date_time,
        activity_civireport.status_id               AS
        civicrm_activity_activity_status_id,
        activity_target_civireport.display_name     AS
        activity_target_civireport_target_contact_id,
        civicrm_activity_target.contact_id          AS target_contact_id,
        activity_assignment_civireport.display_name AS
        activity_assignment_civireport_assignee_contact_id,
        civicrm_activity_assignment.contact_id      AS assignee_contact_id,
        activity_source_civireport.display_name     AS
        activity_source_civireport_source_contact_id,
        civicrm_activity_source.contact_id,
        activity_source_civireport.display_name     AS added_by,
        activity_civireport.activity_date_time      AS date_time_for_sort
 FROM   civicrm_activity activity_civireport
        JOIN civicrm_activity_contact civicrm_activity_source
          ON activity_civireport.id = civicrm_activity_source.activity_id
             AND civicrm_activity_source.record_type_id = 2
             AND civicrm_activity_source.contact_id IN ( 82, 141, 28, 9,
                                                         30, 94, 202, 146,
                                                         92, 158 )
        LEFT JOIN civicrm_activity_contact civicrm_activity_assignment
               ON activity_civireport.id =
                  civicrm_activity_assignment.activity_id
                  AND civicrm_activity_assignment.record_type_id = 1
        LEFT JOIN civicrm_activity_contact civicrm_activity_target
               ON activity_civireport.id = civicrm_activity_target.activity_id
                  AND civicrm_activity_target.record_type_id = 3
        LEFT JOIN civicrm_contact activity_assignment_civireport
               ON civicrm_activity_assignment.contact_id =
                  activity_assignment_civireport.id
        LEFT JOIN civicrm_contact activity_source_civireport
               ON civicrm_activity_source.contact_id =
                  activity_source_civireport.id
        LEFT JOIN civicrm_contact activity_target_civireport
               ON civicrm_activity_target.contact_id =
                  activity_target_civireport.id
        JOIN civicrm_option_value
          ON activity_civireport.activity_type_id = civicrm_option_value.value
             AND ( civicrm_option_value.component_id IS NULL
                    OR civicrm_option_value.component_id IN ( 1, 2, 3, 4,
                                                              6, 8 ) )
        JOIN civicrm_option_group
          ON civicrm_option_group.id = civicrm_option_value.option_group_id
             AND civicrm_option_group.name = 'activity_type'
 WHERE  activity_civireport.is_test = 0)
UNION
(SELECT activity_civireport.id                      AS civicrm_activity_id,
        activity_civireport.activity_type_id        AS
        civicrm_activity_activity_type_id,
        activity_civireport.subject                 AS civicrm_activity_subject,
        activity_civireport.activity_date_time      AS
        civicrm_activity_activity_date_time,
        activity_civireport.status_id               AS
        civicrm_activity_activity_status_id,
        activity_target_civireport.display_name     AS
        activity_target_civireport_target_contact_id,
        civicrm_activity_target.contact_id          AS target_contact_id,
        activity_assignment_civireport.display_name AS
        activity_assignment_civireport_assignee_contact_id,
        civicrm_activity_assignment.contact_id      AS assignee_contact_id,
        activity_source_civireport.display_name     AS
        activity_source_civireport_source_contact_id,
        civicrm_activity_source.contact_id,
        activity_source_civireport.display_name     AS added_by,
        activity_civireport.activity_date_time      AS date_time_for_sort
 FROM   civicrm_activity activity_civireport
        JOIN civicrm_activity_contact civicrm_activity_target
          ON activity_civireport.id = civicrm_activity_target.activity_id
             AND civicrm_activity_target.record_type_id = 3
             AND civicrm_activity_target.contact_id IN ( 82, 141, 28, 9,
                                                         30, 94, 202, 146,
                                                         92, 158 )
        LEFT JOIN civicrm_activity_contact civicrm_activity_assignment
               ON activity_civireport.id =
                  civicrm_activity_assignment.activity_id
                  AND civicrm_activity_assignment.record_type_id = 1
        LEFT JOIN civicrm_activity_contact civicrm_activity_source
               ON activity_civireport.id = civicrm_activity_source.activity_id
                  AND civicrm_activity_source.record_type_id = 2
        LEFT JOIN civicrm_contact activity_assignment_civireport
               ON civicrm_activity_assignment.contact_id =
                  activity_assignment_civireport.id
        LEFT JOIN civicrm_contact activity_source_civireport
               ON civicrm_activity_source.contact_id =
                  activity_source_civireport.id
        LEFT JOIN civicrm_contact activity_target_civireport
               ON civicrm_activity_target.contact_id =
                  activity_target_civireport.id
        JOIN civicrm_option_value
          ON activity_civireport.activity_type_id = civicrm_option_value.value
             AND ( civicrm_option_value.component_id IS NULL
                    OR civicrm_option_value.component_id IN ( 1, 2, 3, 4,
                                                              6, 8 ) )
        JOIN civicrm_option_group
          ON civicrm_option_group.id = civicrm_option_value.option_group_id
             AND civicrm_option_group.name = 'activity_type'
 WHERE  activity_civireport.is_test = 0)
ORDER  BY date_time_for_sort DESC 
```

Comments
----------------------------------------
I also went through and made a bunch of HEREDOCs just to make it easier to work with the SQL.