@eileenmcnaughton 2018-11-09
@agh1 this looks like exactly the sort of reason the activity_contact table was added. Unfortunately when the reports were fixed for it they just replace the activity_assignee & activity_target tables in it. I don't know if that was because the vision got lost by the time it got to the dev or whether it was resource constrained. It was definitely before our PR Review process.

I've been working at reports under CI and there are some tests in api_v3_ReportTemplateTest

Can you see if you can add something that covers activities in this report?