@agh1 2018-11-09
@eileenmcnaughton I'll look at how to test here.  I should note that at present, this PR actually works the same as if there were separate assignee and target tables like there used to be.  The first time I tried this, I ended up just inner joining on the activity_contact table, but that meant that you wouldn't see the other activity contacts on the report.  I tried doing a single inner join against the activity_contact table, but since I still had to do three left joins for the assignee, source, and target, it ended up performing worse than the 4-way union with one fewer join each.