@eileenmcnaughton 2018-11-13
@agh1 I agree that report is pretty weird - it's the only one that does multiple tables on output and I can't think of a sensible way to modify the api to reflect that posibility.

It probably makes sense to test just the function - possibly extracting the activity part out of the clauseComponent function. I think CiviReportTestCase might help set up the objects