@kcristiano 2020-01-26
This PR came up on the last Product development call.   I've run this against a number of my test databases and the before and after results were the same.   The data sets I have are not the same as we see in production, so the improvement in performance is not dramatic.   However, @agh1 has done that work.

I also like the HEREDOCs as the code is more readable and easier to work with.

I agree with @stoob that the hold up is a test.   Andrew  has reviewed this and adding a test has it's own challenges that seem to be more significant than this PR.   Can I suggest that this be merged targeting 5.24 (merged after the 5.23 RC cut) to allow more time for testing.   I don't believe we'll find any issues, but the longer we leave it for testing the better.