@eileenmcnaughton 2020-01-26
@kcristiano I'm OK with merging based on your review. We did discuss this & agreed that we would let it through without a test rather than close it so that's agreed. Ping me if I forget to merge once the next rc is cutt